requirements.txt
# elegantrl
exchange_calendars==3.6.3 # because raise exception with 4.1.1, success tested with 3.6.3
gputil
importlib-metadata==4.13.0
lz4

# plot
matplotlib
ipywidgets

# data handling
numpy
pandas

# Model Building Requirements
scikit-learn>=0.21.0
stable-baselines3
stockstats>=0.4.0
swig
tensorboardX

# market data & paper trading API
yfinance
python-binance
pyfolio
alpaca_trade_api